requirements.txt
requests==2.32.3
cryptography==43.0.3
scipy==1.13.1
python-dotenv==1.0.1
